refactor the code, update unit tests
