+++ src/tests/unit/processors/fixtures/metadata_for_table_with_one_json_column.yaml
