+++ src/tests/unit/processors/fixtures/metadata_for_table_with_two_json_columns.yaml
